wip: refactoring of demo components into rise-ui, part 2
