+++ packages/app/README.md
@@ -1,0 +1,2 @@
+@react-native-templates/app
+====
