+++ packages/app/package.json
@@ -28,6 +28,7 @@
   "homepage": "https://github.com/ericvicenti/react-native-templates#readme",
   "dependencies": {
     "@hookform/resolvers": "^3.3.4",
+    "@react-native-templates/ui-tamagui": "0.1.0",
     "@react-native-templates/core": "0.1.0",
     "@react-native-templates/demo-ui": "0.0.1",
     "@react-native-templates/ws-client": "0.1.0",
