+++ packages/app/tsconfig.json
@@ -13,6 +13,9 @@
     },
     {
       "path": "../ws-client"
+    },
+    {
+      "path": "../ui-tamagui"
     }
   ]
 }
