+++ packages/ui-rise/README.md
@@ -1,0 +1,2 @@
+@react-native-templates/ui-rise
+=====
