+++ packages/ui-rise/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@react-native-templates/ui-tamagui",
-  "version": "0.1.0",
+  "name": "@react-native-templates/ui-rise",
+  "version": "0.0.1",
   "license": "Apache-2.0",
   "description": "",
   "main": "lib",
@@ -16,7 +16,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/ericvicenti/react-native-templates.git",
-    "directory": "packages/ui-tamagui"
+    "directory": "packages/ui-rise"
   },
   "author": "Mike Grabowski <grabbou@gmail.com> (https://github.com/grabbou)",
   "bugs": {
@@ -29,6 +29,14 @@
     "@types/react": "^18.0.27",
     "react": "^18.2.0",
     "tamagui": "^1.94.4"
+  },
+  "dependencies": {
+    "@tamagui/linear-gradient": "^1.94.4",
+    "@tamagui/lucide-icons": "^1.94.4",
+    "react-native-draggable-flatlist": "^4.0.1",
+    "react-native-gesture-handler": "^2.16.0",
+    "react-native-qrcode-svg": "^6.3.0",
+    "zod": "^3.22.4"
   },
   "peerDependencies": {
     "react": "^18.2.0",
