+++ packages/ui-rise/src/QRCode.tsx
@@ -1,0 +1,19 @@
+import { ComponentProps } from '@react-native-templates/core'
+import React from 'react'
+import RNQRCode from 'react-native-qrcode-svg'
+import { Spinner } from 'tamagui'
+import { z } from 'zod'
+
+const QRCodeProps = z.object({
+  value: z.string().nullable().optional(),
+})
+
+// tbd: provide web support
+export function QRCode(props: z.infer<typeof QRCodeProps> & ComponentProps) {
+  if (!props.value) return <Spinner />
+  return <RNQRCode color="white" backgroundColor="black" value={props.value} />
+}
+
+QRCode.validate = (props: any) => {
+  return QRCodeProps.parse(props)
+}
