+++ packages/ui-tamagui/src/index.tsx
@@ -1,4 +1,4 @@
-import { ComponentRegistry } from '@react-native-templates/core'
+import { ComponentRegistry, wrapEvents } from '@react-native-templates/core'
 import * as t from 'tamagui'
 
 const TamaguiComponents: ComponentRegistry = {
@@ -49,7 +49,7 @@ const TamaguiComponents: ComponentRegistry = {
   },
   /* https://tamagui.dev/ui/button */
   Button: {
-    component: t.Button,
+    component: wrapEvents(t.Button, ['onPress', 'onPressIn', 'onPressOut']),
   },
   /* https://tamagui.dev/ui/checkbox */
   Checkbox: {
