+++ tsconfig.json
@@ -6,7 +6,8 @@
       "@react-native-templates/demo-ui": ["./packages/demo-ui/src"],
       "@react-native-templates/core": ["./packages/core/src"],
       "@react-native-templates/ws-client": ["./packages/ws-client/src"],
-      "@react-native-templates/ui-tamagui": ["./packages/ui-tamagui/src"]
+      "@react-native-templates/ui-tamagui": ["./packages/ui-tamagui/src"],
+      "@react-native-templates/ui-rise": ["./packages/ui-rise/src"]
     },
     "composite": true,
     "strictNullChecks": true,
